requirements.txt
aiomysql==0.0.20
alembic==1.4.1
autopep8==1.5
bcrypt==3.1.7
cffi==1.14.0
Click==7.0
cryptography==2.8
databases==0.2.6
fastapi==0.52.0
h11==0.9.0
httptools==0.1.1
Mako==1.1.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
orm==0.1.5
passlib==1.7.2
pkg-resources==0.0.0
pycodestyle==2.5.0
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
PyMySQL==0.9.2
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-multipart==0.0.5
six==1.14.0
SQLAlchemy==1.3.13
starlette==0.13.2
typesystem==0.2.4
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
